cdk/lambda/writeback-handler/requirements.txt
psycopg2-binary==2.9.6
requests
